Wire General Assist and screen capture through Electron IPC
