+++ surfsense_desktop/src/modules/tray.ts
@@ -1,31 +1,20 @@
-import { app, globalShortcut, Menu, nativeImage, Tray } from 'electron';
+import { app, globalShortcut, Menu, nativeImage, Tray, type NativeImage } from 'electron';
 import path from 'path';
-import { getMainWindow, createMainWindow } from './window';
+import { runGeneralAssistShortcut } from './general-assist';
+import { showMainWindow } from './window';
 import { getShortcuts } from './shortcuts';
 import { trackEvent } from './analytics';
 
 let tray: Tray | null = null;
 let currentShortcut: string | null = null;
 
-function getTrayIcon(): nativeImage {
+function getTrayIcon(): NativeImage {
   const iconName = process.platform === 'win32' ? 'icon.ico' : 'icon.png';
   const iconPath = app.isPackaged
     ? path.join(process.resourcesPath, 'assets', iconName)
     : path.join(__dirname, '..', 'assets', iconName);
   const img = nativeImage.createFromPath(iconPath);
   return img.resize({ width: 16, height: 16 });
-}
-
-function showMainWindow(source: 'tray_click' | 'tray_menu' | 'shortcut' = 'tray_click'): void {
-  const existing = getMainWindow();
-  const reopened = !existing || existing.isDestroyed();
-  if (reopened) {
-    createMainWindow('/dashboard');
-  } else {
-    existing.show();
-    existing.focus();
-  }
-  trackEvent('desktop_main_window_shown', { source, reopened });
 }
 
 function registerShortcut(accelerator: string): void {
@@ -35,11 +24,14 @@ function registerShortcut(accelerator: string): void {
   }
   if (!accelerator) return;
   try {
-    const ok = globalShortcut.register(accelerator, () => showMainWindow('shortcut'));
+    const ok = globalShortcut.register(accelerator, () => {
+      void runGeneralAssistShortcut();
+    });
     if (ok) {
       currentShortcut = accelerator;
+      console.log(`[general-assist] Register ${accelerator}: OK`);
     } else {
-      console.warn(`[tray] Failed to register General Assist shortcut: ${accelerator}`);
+      console.warn(`[general-assist] Register ${accelerator}: FAILED (OS or another app may own this chord)`);
     }
   } catch (err) {
     console.error(`[tray] Error registering General Assist shortcut:`, err);
